Enemy and cutting testing
PROBLEM: Centerpoint calculation is not optimal. On slices that curve
inside to the shape, the curve side gets a sharp edge and the mesh is
somewhat turned in z-axis.

PROBLEM: Original mesh loses a vertice on every cut. The vertice is at
the top and it is always the next from the last lost. Problem in the
adding of vertices on creation of triangles.

Otherwise no new things.